Refactor: create files for each handler
